Refactor mission-related data fetching to use custom hooks
This commit refactors the way missions data is fetched by introducing the use of custom hooks and context. It removes the old method of calling API directly within the MissionTable component and instead brings in the useMissions hook to handle the request. A new MissionContext is also added to hold shared mission-related data and operations. Code formatting in several files and responses handling from apiService has been adjusted as well.