Update default values and adjust styles in various components
This commit updates the default values of the parameters in the EclipseTimeForm and EclipseTimePage components. The default body has been changed to "Earth" and the default apoapsis and periapsis have been adjusted to "1000". Some style changes have been made to the HomePage and CelestialBody components, specifically adjusting the height of the canvas container and the sphereGeometry arguments respectively.