Add password visibility toggle in LoginForm
Introduced a new feature in LoginForm where the user can toggle the visibility of their input password. This was done by creating state for password visibility and inserting a button to handle the toggling. Also, updated the password requirement in the formSchema from minimum length of 8 characters to simply being required.